requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
deepface==0.0.93
opencv-python==4.11.0.86
numpy==2.1.3
tensorflow==2.19.0
Pillow==11.2.1
gunicorn==23.0.0
python-decouple==3.8
paddlepaddle==2.6.0
paddleocr==2.7.0.3
transformers==4.37.2
torch==2.1.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0 